JTable Not handling default values correctly in MSSQL 2008+ (#11056)
* JTable Not handling default values correctly in MSSQL 2008+

Default values are not being assigned correctly for data types such as
Integer, Date and Datetime. So as a result it is causing the insert
query to fail. I have tested this on SQL SERVER 2008-2016.

The root cause of this issue lies with the getTableColumns in the SQL
server driver class, which relies on using the information_schema views
to get the information about the columns. However the default values it
gets for the table columns are from the underlying system comments table
and as a result it is not in a format that can be injected as is into a
INSERT Query.

Below is a partial list of how the syscomments table stores default
values for the following data types:

Datatype              Actual default Value                 Value  in
syscomments table

Int                           0
((0))
Int                          10
((1))
DataTime             ‘1900-01-01 00:00:00’                (‘1900-01-01
00:00:00’)
DataTime             ‘1900-01-01T00:00:00.000’
(‘1900-01-01T00:00:00.000’)
DataTime             getdate()
(getdate())
Date                   ‘1900-01-01’
(‘1900-01-01’)
Varchar              ‘’
(’’)
Varchar             ‘1234’
(‘1234’)
Nvarchar             N‘’
(N’’)
Nvarchar             N‘1234’
(N‘1234’)
Nvarchar             NULL
(NULL)

You see the general idea that the default value stored in the
syscomments table is wrapped around in curly brackets, whether that be
single or double depends on the data type.

Now the current implementation of the getTableColumns function only
deals with nvarchar  data types by assuming every default string value
will be an empty string  so defaults to that value.  This works but IMHO
is a bit too simplistic in that it does not allow for default string
values that is not an empty string value.  Also it does not deal with
other data types such as datetime, date etc... So as a result the INSERT
query will blow up when you try for example to pass the value ‘((0))’
for an integer data type.

My change just parses the value by removing the superfluous brackets and
returns the actual value for the data type in the correct format.

So for example the value ‘((0))’ will be returned as ‘0’ and (N‘1’) will
be returned as ‘1’.

* Forgot to add rule for protecting default function computed values